service/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
azure-core==1.31.0
azure-identity==1.19.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
environs==11.0.0
fastapi==0.115.0
fastapi_cors==0.0.6
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
llama-index-core==0.11.18
llama-index-graph-stores-neo4j==0.3.3
llama-index-llms-azure-openai==0.2.2
llama-index-llms-openai==0.2.15
llama-index-tools-openapi==0.2.0
llama-index-vector-stores-neo4jvector==0.2.3
MarkupSafe==3.0.1
marshmallow==3.22.0
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
mypy==1.11.2
mypy-extensions==1.0.0
neo4j==5.25.0
nest-asyncio==1.6.0
networkx==3.4.1
nltk==3.9.1
numpy==1.26.4
openai==1.51.2
packaging==24.1
pillow==10.4.0
portalocker==2.10.1
prometheus_client==0.21.0
prometheus_flask_exporter==0.23.1
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tenacity==8.5.0
tiktoken==0.8.0
tqdm==4.66.5
types-PyYAML==6.0.12.20240917
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.15.2
